# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: progress
    description: User progress persistence
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/user:
    get:
      operationId: getCurrentAuthUser
      tags: [auth]
      summary: Get current authenticated user
      description: Returns the current auth state. Returns null user when not authenticated.
      responses:
        "200":
          description: Current auth state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCurrentAuthUserResponse"

  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile authorization code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"

  /mobile-auth/logout:
    post:
      operationId: logoutMobileSession
      tags: [auth]
      summary: Logout mobile session
      responses:
        "200":
          description: Logout success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutMobileSessionResponse"

  /progress:
    get:
      operationId: getUserProgress
      tags: [progress]
      summary: Fetch the current user's saved progress
      description: Returns the user's AppState blob. Returns null data when no progress saved yet.
      responses:
        "200":
          description: User progress data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserProgressResponse"
        "401":
          description: Not authenticated
    put:
      operationId: saveUserProgress
      tags: [progress]
      summary: Save the current user's progress
      description: Upserts the full AppState blob for the authenticated user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveUserProgressBody"
      responses:
        "200":
          description: Save confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveUserProgressResponse"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
      required:
        - id

    GetCurrentAuthUserResponse:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required:
        - user

    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required:
        - code
        - code_verifier
        - redirect_uri
        - state

    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token

    LogoutMobileSessionResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    GetUserProgressResponse:
      type: object
      properties:
        data:
          nullable: true
      required:
        - data

    SaveUserProgressBody:
      type: object
      properties:
        data:
          type: object
      required:
        - data

    SaveUserProgressResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success