# lib/api-zod/src/generated/types/saveUserProgressBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SaveUserProgressBodyData } from "./saveUserProgressBodyData";

export interface SaveUserProgressBody {
  data: SaveUserProgressBodyData;
}